app/requirements.txt
tuya-connector-python
tuya-iot-py-sdk
python-dotenv
duckdb
pandas
